feat(semantic): add node type resolving
