+++ crates/toto_tosca/src/grammar/v2_0/mod.rs
@@ -73,5 +73,7 @@ mod tests {
         dbg!(Dot::new(&ast));
 
         get_errors(&ast).for_each(|(what, loc)| report_error(what, loc, &ast));
+
+        assert!(false);
     }
 }
